// types.ts
export interface LessonPlanData {
  schoolName: string;
  teacherName: string;
  subject: string;
  phaseClass: string;
  timeAllocation: string;
  schoolYear: string;
  studentProfile: string;
  subjectMatter: string;
  specificMaterial?: string;
  graduateProfileDimensions: string[];
  learningOutcomes: string;
  crossDiscipline: string;
  learningObjectives: string;
  learningTopic: string;
  pedagogicalPractices: string;
  learningPartnership: string;
  learningEnvironment: string;
  digitalUtilization: string;
  initialActivities: string;
  coreActivities: string;
  closingActivities: string;
  initialAssessment: string;
  processAssessment: string;
  finalAssessment: string;
  studentWorksheet: string;
  signatureDate: string;
  principalName: string;
  principalNip: string;
  teacherNip: string;
}